Added Cart System With Spring Security

+++ Spring Boot REST API/src/main/java/com/sagar/dao/UserRepository.java
@@ -5,9 +5,10 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
 import com.sagar.entity.User;
- 
-public interface UserRepository extends CrudRepository<User, Long> {
- 
-    @Query("SELECT u FROM User u WHERE u.username = :username")
-    public User getUserByUsername(@Param("username") String username);
+
+public interface UserRepository extends CrudRepository<User, Integer> {
+
+	@Query("SELECT u FROM User u WHERE u.username = :username")
+	public User getUserByUsername(@Param("username") String username);
+
 }
